movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
factory-boy==3.2.0